Refactor WebsiteDataSearchResultViewController with ThemedTableViewController, Refactor WebsiteDataManagementViewController and WebsiteDataSearchResultViewController with dequeue reusable cell for indexPath.
